msm: camera: Fix memory leak
CAF never deattaches the dma buffer, creating a memory leak
every time this function is run. This issue becomes increasingly
problematic as it's run a metric ton every time camera is open,
which on Coral is pretty much every time you unlock the phone
and often while you run it too. Fix it by unmaping and detaching
the dma buffer.

Reported-by: celtare21 <celtare21@gmail.com>
Signed-off-by: Demon000 <demonsingur@gmail.com>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: Edgars Cīrulis <edgarsscirulis@gmail.com>